Add custom Ord instance for Package.
This lets us avoid duplicates which may occur if there are multiple packages
with the same under a different attribute path, such as xen, xen-light,
xen-slim. Without this change they would be considered unique packages and a
single CVE would be reported for all which is noisy.

This should generally be safe although it may lead to issue with packages such
as libxml2 and pythonPackages.libxml2 which are now considered the same. One
option to avoid it would be to collapse the attribute path and normalise the
name of the latter to pythonPackages-libxml2.