nvs: switch over to parsing store derivations over Nix expressions
Instead of parsing a nixpkgs tree, or any other expression evaluating to an attribute set of
packages, use the .drv store derivation to query packages to be scanned. This allows for example
checking the current running system easily as well as querying the derivations for already applied
CVE patches (although this is merely by convention).