nvs: fix up package name & version parsing
The implementations did not actually conform with Nix's `parseDrvName`. For example parsing package
name out of `hello` (note no version string at all) would result in an empty name, the same would
occur for `hello-`. In both of those cases the name ought to be the entirety of the string.